+++ sdk/trace/src/test/java/io/opentelemetry/sdk/trace/RandomIdGeneratorTest.java
@@ -26,18 +26,4 @@ class RandomIdGeneratorTest {
       assertThat(spanId).isNotEqualTo(SpanId.getInvalid());
     }
   }
-
-  @Test
-  void androidVersion() {
-    IdGenerator generator = AndroidFriendlyRandomIdGenerator.INSTANCE;
-
-    // Can't assert values but can assert they're valid, try a lot as a sort of fuzz check.
-    for (int i = 0; i < 1000; i++) {
-      String traceId = generator.generateTraceId();
-      assertThat(traceId).isNotEqualTo(TraceId.getInvalid());
-
-      String spanId = generator.generateSpanId();
-      assertThat(spanId).isNotEqualTo(SpanId.getInvalid());
-    }
-  }
 }
